Cli create command update (#31)
* feat(gazelle_cli): Update to `gazelle_core 0.3.0`

Updated `create` command to use `gazelle_core 0.3.0`

* feat(gazelle_cli): easier `create` command

The `create` command is now much easier to use.
It now asks for a project name, and the developer does not need
to specify a path or a name with options.

BREAKING CHANGE: -n and -p options have been removed

* feat(gazelle_cli): Create handler

Created a `createHandler` function that generates a new handler
for a given route. Refactored `createProject` function.

* feat(gazelle_cli): CreateRoute command

* feat(gazelle_cli): CreateHook function

* feat(gazelle_cli): WIP - CreateProject

Started to udpate the `createProject` function to
scaffold a whole project with routes, handlers and hooks.

* feat(gazelle_cli): CreateProject command

Now `gazelle create` creates a project with a stronger
structure.

* feat(gazelle_cli): CreateProject && CreateRoute

Added `CreateProject` command and `CreateRoute` command.

* fix(gazelle_cli): Removed unnecessary import

* feat(gazelle_cli): Create command input cleaning

Create commands now sanitize string input to be compliant
with dart naming conventions.

* fix(gazelle_cli): Wrong entrypoint import

* feat(gazelle_cli): Gazelle project detection

`loadProjectConfiguration` now also evaluates if the
current folders actually containes a Gazelle project.
When a Gazelle project isn't detected, an exception is
thrown.Gazelle project are detected by using the `gazelle`
property inside the pubspec.yaml file.