feat(gazelle_core): Serialization (#35)
* feat(gazelle_core) WIP: GazelleModelType

Created a `GazelleModelType` class to handle serialization
and deserialization under the hood.

* feat(gazelle_core): WIP - Response serialization

Handled primitive types, Lists and custom types.
Still need to handle Maps.

* feat(gazelle_core): WIP - Response serialization

Finished to implement the serialization function with relative tests.
The CLI will generate models based on this feature.
Still need to attach this to the actual server.

* feat(gazelle_core): Serialization

Attached new serialization function to the actual HTTP server.

* feat(gazelle_core): Model Provider reference

Added `GazelleModelProvider` reference in `GazelleApp`
constructor. Once a ModelProvider is written/generated
it can be passed to a `GazelleApp` to serialize/deserialize
registered types. As of now, this is manual work from
a developer, but it will be automatically generated
by the CLI soon.

* feat(gazelle_core): Automatic content/tpe header

The content-type header is now automatically set by Gazelle
based on the type of `GazelleResponse.body`.