feat(gazelle_core): Export route structure
This features enables the router to export its structure
to a json map. This is extremely helpful for client
generation as it is not possible to get the full structure
of the router just by looking at project files as
plugins can also add routes to the server in isolation,
thus, we need to populate the router and export it
at runtime.
The CLI will pick this output and use it to generate
the necessary code to have a complete client.
To use this feature, the sever will be launched with
a flag that populates the router with all the needed
routes and exports them, but the server never listens
to any incoming connections during this phase.