feat(gazelle_core): Type system
Updated the type system to have better type safety
both on the server and client.

BREAKING CHANGE: Handlers are now defined with a
class called `GazelleRouteHandler`.
This changes gives better type safety and structure
to a Gazelle project. It also enables us to
extend handlers with ease in the future.