feat(gazelle_core): Client generation (#49)
* feat(gazelle_core): Export route structure

This features enables the router to export its structure
to a json map. This is extremely helpful for client
generation as it is not possible to get the full structure
of the router just by looking at project files as
plugins can also add routes to the server in isolation,
thus, we need to populate the router and export it
at runtime.
The CLI will pick this output and use it to generate
the necessary code to have a complete client.
To use this feature, the sever will be launched with
a flag that populates the router with all the needed
routes and exports them, but the server never listens
to any incoming connections during this phase.

* [WIP] feat(gazelle_core): Export routes

Added basic functionality to start a `GazelleApp` and
export its routes. Still need to finish testing
and catch the result with the CLI.

* feat(gazelle_cli): Basic client generation

The CLI now generates a client usable by other
dart projects. It still needs a couple of things
to make everything good DX wise.

* feat(gazelle_cli): Client generation

Client generation is almost done.
Still need to handle model types properly inside
extension methods, other than that, is pretty
much finished.

* fix(gazelle_core): Fixed tests

Fixed broken test for `GazelleRouter`.

* [WIP] feat(gazelle_core): return types

Started working on better return types
to help developers with static type
inference inside route handlers.

* [WIP] feat(gazelle_core): Type system

* feat(gazelle_core): Type system

Updated the type system to have better type safety
both on the server and client.

BREAKING CHANGE: Handlers are now defined with a
class called `GazelleRouteHandler`.
This changes gives better type safety and structure
to a Gazelle project. It also enables us to
extend handlers with ease in the future.

* feat(gazelle_core): Type system

Updated the type system to have better type safety
both on the server and client.

BREAKING CHANGE: Handlers are now defined with a
class called `GazelleRouteHandler`.
This changes gives better type safety and structure
to a Gazelle project. It also enables us to
extend handlers with ease in the future.

* refactor(plugins): Updated plugins

Updated plugins with new `gazelle_core` interface.

* feat(gazelle_cli): Get latest package version

Implemented `getLatestPackageVersion` to get the latest
version of a given package.
This is really helpful inside the CLI as we do not need
to update code generation parts whenever a gazelle
package gets updated.