feat(gazelle_core): New handler interface
Converted `GazelleRouteHandler` to an abstract class.
This helps a lot with type safety and return types both
for the server and client. It adds a bit of verbosity,
but it is much easier to maintain and we can extend this
with a lot of features in the future. The CLI will help
developers create new handlers with ease, although anyone
can just create a class that extends `GazelleRouteHandler`
and overrides the `call()` method.

BREAKING CHANGE: `GazelleRouteHandler` is not a typdef of
a function anymore. It is now an abstract class that needs
to be extended and implemented.