feat(gazelle_cli): Create project update
Create project now sets up also an empty model
provider. This is helpful as the developer
does not need to import the model provdier in the server
and/or client. From now on, this should be handled
entirely by the CLI during the creation process.
The developer can still edit every file as desired,
but it shouldn't be needed.