feat: Developer Experience (#50)
* feat(gazelle_core): New handler interface

Converted `GazelleRouteHandler` to an abstract class.
This helps a lot with type safety and return types both
for the server and client. It adds a bit of verbosity,
but it is much easier to maintain and we can extend this
with a lot of features in the future. The CLI will help
developers create new handlers with ease, although anyone
can just create a class that extends `GazelleRouteHandler`
and overrides the `call()` method.

BREAKING CHANGE: `GazelleRouteHandler` is not a typdef of
a function anymore. It is now an abstract class that needs
to be extended and implemented.

* feat(plugins): Updated with core breaking changes

Updated all of the plugins with the latest `gazelle_core`
breaking changes.

* feat(gazelle_cli): Updated create handler

Updated create handler command with the new
`GazelleRouteHandler` interface.

* feat(gazelle_cli): Create project update

Create project now sets up also an empty model
provider. This is helpful as the developer
does not need to import the model provdier in the server
and/or client. From now on, this should be handled
entirely by the CLI during the creation process.
The developer can still edit every file as desired,
but it shouldn't be needed.

* feat(gazelle_cli): Better project setup

Enhanced create project command and client generation.

* fix(gazelle_client): Fixed deserialization

* fix(gazelle_cli): Run command

Fixed a bug that made the `run` command
throw an error when launched outside
the server directory.

* [WIP] feat(gazelle_cli): Flutter project

* [WIP] feat(gazelle_cli): Flutter project

* feat(gazelle_cli): Flutter project

`gazelle create project --full-stack` now
scaffolds a full-stack project with Flutter.
Backend and frontend are already linked.

* fix(gazelle_core): Fixed default port setting

* feat(gazelle_core): Changed create option

Changed create option to `--flutter`.