feat(gazelle_core): Handler interface
Updated the way to define a handler.
Handlers are now a simple function, and Gazelle
will do the serialziation and deserialization
of requests and responses under the hood,
so the body of the request and response is typed.

BREAKING CHANGE: The `GazelleHandler` interface
has been removed in favor of a simple top-level
function to make it easier to define handlers.
The handlers are attached to the route with
a builder pattern instead of constructor
parameters.